fastapi/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
xgboost
google-genai
python-dotenv